* Inject Tileo/TileAtlasBuilder into AllegroFlare/TileMaps/PrimMeshAtlas
