+++ documentation/index.htm
@@ -9740,9 +9740,6 @@ table td
   <td class="dependency">{&quot;symbol&quot;=&gt;&quot;AllegroFlare::TileMaps::TileMap&quot;, &quot;headers&quot;=&gt;[&quot;AllegroFlare/TileMaps/TileMap.hpp&quot;]}</td>
 </tr>
 <tr>
-  <td class="dependency">{&quot;symbol&quot;=&gt;&quot;Tileo::TileAtlasBuilder&quot;, &quot;headers&quot;=&gt;[&quot;Tileo/TileAtlasBuilder.hpp&quot;]}</td>
-</tr>
-<tr>
   <td class="dependency">{&quot;symbol&quot;=&gt;&quot;Tileo::TMJDataLoader&quot;, &quot;headers&quot;=&gt;[&quot;Tileo/TMJDataLoader.hpp&quot;]}</td>
 </tr>
     </table>
@@ -11458,9 +11455,6 @@ table td
   <td class="dependency">{&quot;symbol&quot;=&gt;&quot;std::vector&lt;Wicked::Entities::Basic2D*&gt;&quot;, &quot;headers&quot;=&gt;[&quot;vector&quot;, &quot;Wicked/Entities/Basic2D.hpp&quot;]}</td>
 </tr>
 <tr>
-  <td class="dependency">{&quot;symbol&quot;=&gt;&quot;TileAtlasBuilder&quot;, &quot;headers&quot;=&gt;[&quot;Tileo/TileAtlasBuilder.hpp&quot;]}</td>
-</tr>
-<tr>
   <td class="dependency">{&quot;symbol&quot;=&gt;&quot;Wicked::Entities::Basic2D*&quot;, &quot;headers&quot;=&gt;[&quot;Wicked/Entities/Basic2D.hpp&quot;]}</td>
 </tr>
 <tr>
@@ -13067,9 +13061,6 @@ table td
   "ifstream": [
     "Tileo/TMJMeshLoader"
   ],
-  "Tileo/TileAtlasBuilder": [
-    "Tileo/TMJMeshLoader"
-  ],
   "Tileo/TMJDataLoader": [
     "Tileo/TMJMeshLoader"
   ],
@@ -13217,9 +13208,6 @@ table td
     "WickedDemos/TileDemo"
   ],
   "std/vector<Wicked/Entities/Basic2D>": [
-    "WickedDemos/TileDemo"
-  ],
-  "TileAtlasBuilder": [
     "WickedDemos/TileDemo"
   ],
   "Wicked/Entities/Basic2DFactory": [
