+++ include/AllegroFlare/TileMaps/PrimMeshAtlas.hpp
@@ -49,6 +49,39 @@ namespace AllegroFlare::TileMaps
       bool get_tile_uv(int index_num, float *u1, float *v1, float *u2, float *v2);
 
       std::vector<AllegroFlare::TileMaps::PrimMeshAtlasIndexRecord> get_tile_index();
+
+
+
+
+
+
+class TileAtlasBuilder
+{
+private:
+   int tile_w;
+   int tile_h;
+   std::vector<AllegroFlare::TileMaps::PrimMeshAtlasIndexRecord> tile_index;
+
+public:
+   TileAtlasBuilder(
+      int tile_w=0,
+      int tile_h=0,
+      std::vector<AllegroFlare::TileMaps::PrimMeshAtlasIndexRecord> tile_index={}
+   );
+   ~TileAtlasBuilder();
+
+   ALLEGRO_BITMAP *build_extruded();
+   static ALLEGRO_BITMAP *build_scaled_and_extruded(ALLEGRO_BITMAP *bitmap=nullptr, int scale=3);
+   static ALLEGRO_BITMAP *create_pixel_perfect_scaled_render(ALLEGRO_BITMAP *bitmap=nullptr, int scale=4);
+};
+
+
+
+
+   // brought over from Tileo/TileAtlasBuilder for refactoring purposes
+      //static ALLEGRO_BITMAP *build_scaled_and_extruded(ALLEGRO_BITMAP *bitmap=nullptr, int scale=3);
+      //static ALLEGRO_BITMAP *create_pixel_perfect_scaled_render(ALLEGRO_BITMAP *bitmap=nullptr, int scale=4);
+
    };
 }
